+++ src/installers/pipPackage.ts
@@ -1,0 +1,3 @@
+export default class PipPackage {
+    constructor(public name: string, public version: string) { }
+}