+++ src/installers/poetry.ts
@@ -1,32 +1,31 @@
-import exec from '../exec';
-import { getPythonPath, getWorkDir } from '../pythonPath';
 import * as path from 'path';
 import * as fs from 'fs';
+import { BaseInstaller } from './base';
 
-export function isPresent(): boolean {
-    return fs.existsSync(path.join(getWorkDir(), 'poetry.lock'));
-}
 
-export async function installPackage(name: string, version?: string, extras: string[] = [], dev: boolean = true): Promise<void> {
-    let nameExtras = name;
-    if (extras.length) {
-        const extrasStr = extras.join(',');
-        nameExtras = `${nameExtras}[${extrasStr}]`;
+export default class PoetryInstaller extends BaseInstaller {
+    async installPackage(name: string, version: string, extras: string[], dev: boolean): Promise<void> {
+        const packageName = this.buildPackageName(name, extras);
+        const versionConstraint = version ? this.buildVersionConstraint(version) : '@latest';
+        const cmd = `${this.pythonPath} -m poetry add -n ${dev ? '--dev' : ''} "${packageName}${versionConstraint}"`;
+        await this.exec(cmd);
     }
-    const versionConstraint = version ? `${(version.startsWith('=') || version.startsWith('<') || version.startsWith('>')) ? '' : '=='}${version}` : '@latest';
-    const cmd = `${getPythonPath()} -m poetry add -n ${dev ? '--dev' : ''} "${nameExtras}${versionConstraint}"`;
-    console.log(cmd);
-    const oldCwd = process.cwd();
-    process.chdir(getWorkDir());
-    await exec(cmd);
-    process.chdir(oldCwd);
-}
 
-export async function removePackage(name: string) {
-    const cmd = `${getPythonPath()} -m poetry remove -n "${name}"`;
-    console.log(cmd);
-    const oldCwd = process.cwd();
-    process.chdir(getWorkDir());
-    await exec(cmd);
-    process.chdir(oldCwd);
+    async removePackage(name: string, dev: boolean): Promise<void> {
+        const cmd = `${this.pythonPath} -m poetry remove -n ${dev ? '--dev' : ''} "${name}"`;
+        await this.exec(cmd);
+    }
+
+    isInUse(): boolean {
+        if (!this.lockFileExists()) { return false; }
+        const lockFilePath = path.join(this.workDir, 'poetry.lock');
+        const data = fs.readFileSync(lockFilePath, { encoding: 'utf-8' });
+        if (data.includes('"mypy-boto3')) { return true; }
+        if (data.includes('"boto3-stubs"')) { return true; }
+        return false;
+    }
+
+    lockFileExists(): boolean {
+        return fs.existsSync(path.join(this.workDir, 'poetry.lock'));
+    }
 }