+++ CHANGELOG.md
@@ -2,12 +2,17 @@
 
 All notable changes to the "boto3-ide" extension will be documented in this file.
 
-## [Unreleased]
+## [0.2.0]
 ### Added
 - `pipenv` support
 
+### Changed
+- Packages are installed with `poetry` if they have been installed with it previously
+- Installer allows to choose between `poetry` and `pip` if `poetry` is not used yet for installation
+
 ### Fixed
 - `poetry` correctly removes master module if it is not needed
+- Services autodiscovery did not check all folders
 
 ## [0.1.0]